add cpp_only fbe to fix golang ci
