+++ proto/cpp_only.fbe
@@ -1,0 +1,10 @@
+package cpp_only
+
+// int128 and uint128
+struct Struct128
+{
+    int128 f1;
+    int128? f2;
+    uint128 f3;
+    uint128? f4;
+}
