+++ proto/cpp_only_final_protocol.cpp
@@ -1,0 +1,45 @@
+//------------------------------------------------------------------------------
+// Automatically generated by the Fast Binary Encoding compiler, do not modify!
+// https://github.com/chronoxor/FastBinaryEncoding
+// Source: cpp_only.fbe
+// FBE version: 1.11.0.0
+//------------------------------------------------------------------------------
+
+#if defined(_MSC_VER)
+#pragma warning(push)
+#pragma warning(disable:4065) // C4065: switch statement contains 'default' but no 'case' labels
+#endif
+
+#include "cpp_only_final_protocol.h"
+
+namespace FBE {
+
+namespace cpp_only {
+
+bool FinalReceiver::onReceive(size_t type, const void* data, size_t size)
+{
+    switch (type)
+    {
+        default: break;
+    }
+
+    return false;
+}
+
+void FinalClient::reset_requests()
+{
+    Sender::reset();
+    Receiver::reset();
+}
+
+void FinalClient::watchdog_requests(uint64_t utc)
+{
+}
+
+} // namespace cpp_only
+
+} // namespace FBE
+
+#if defined(_MSC_VER)
+#pragma warning(pop)
+#endif
