+++ proto/test.h
@@ -862,54 +862,4 @@ struct std::hash<test::StructEmpty>
 
 namespace test {
 
-struct Struct128
-{
-    __int128_t f1;
-    std::optional<__int128_t> f2;
-    __uint128_t f3;
-    std::optional<__uint128_t> f4;
-
-    size_t fbe_type() const noexcept { return 144; }
-
-    Struct128();
-    Struct128(__int128_t arg_f1, const std::optional<__int128_t>& arg_f2, __uint128_t arg_f3, const std::optional<__uint128_t>& arg_f4);
-    Struct128(const Struct128& other) = default;
-    Struct128(Struct128&& other) = default;
-    ~Struct128() = default;
-
-    Struct128& operator=(const Struct128& other) = default;
-    Struct128& operator=(Struct128&& other) = default;
-
-    bool operator==(const Struct128& other) const noexcept;
-    bool operator!=(const Struct128& other) const noexcept { return !operator==(other); }
-    bool operator<(const Struct128& other) const noexcept;
-    bool operator<=(const Struct128& other) const noexcept { return operator<(other) || operator==(other); }
-    bool operator>(const Struct128& other) const noexcept { return !operator<=(other); }
-    bool operator>=(const Struct128& other) const noexcept { return !operator<(other); }
-
-    std::string string() const;
-
-    friend std::ostream& operator<<(std::ostream& stream, [[maybe_unused]] const Struct128& value);
-
-    void swap(Struct128& other) noexcept;
-    friend void swap(Struct128& value1, Struct128& value2) noexcept { value1.swap(value2); }
-};
-
-} // namespace test
-
-template<>
-struct std::hash<test::Struct128>
-{
-    typedef test::Struct128 argument_type;
-    typedef size_t result_type;
-
-    result_type operator() ([[maybe_unused]] const argument_type& value) const
-    {
-        result_type result = 17;
-        return result;
-    }
-};
-
-namespace test {
-
 } // namespace test
