+++ proto/test_final_models.h
@@ -1123,80 +1123,4 @@ private:
 
 } // namespace test
 
-// Fast Binary Encoding ::test::Struct128 final model
-template <>
-class FinalModel<::test::Struct128>
-{
-public:
-    FinalModel(FBEBuffer& buffer, size_t offset) noexcept;
-
-    // Get the allocation size
-    size_t fbe_allocation_size(const ::test::Struct128& fbe_value) const noexcept;
-    // Get the final offset
-    size_t fbe_offset() const noexcept { return _offset; }
-    // Set the final offset
-    size_t fbe_offset(size_t offset) const noexcept { return _offset = offset; }
-    // Get the final type
-    static constexpr size_t fbe_type() noexcept { return 144; }
-
-    // Shift the current final offset
-    void fbe_shift(size_t size) noexcept { _offset += size; }
-    // Unshift the current final offset
-    void fbe_unshift(size_t size) noexcept { _offset -= size; }
-
-    // Check if the struct value is valid
-    size_t verify() const noexcept;
-    // Check if the struct fields are valid
-    size_t verify_fields() const noexcept;
-
-    // Get the struct value
-    size_t get(::test::Struct128& fbe_value) const noexcept;
-    // Get the struct fields values
-    size_t get_fields(::test::Struct128& fbe_value) const noexcept;
-
-    // Set the struct value
-    size_t set(const ::test::Struct128& fbe_value) noexcept;
-    // Set the struct fields values
-    size_t set_fields(const ::test::Struct128& fbe_value) noexcept;
-
-private:
-    FBEBuffer& _buffer;
-    mutable size_t _offset;
-
-public:
-    FinalModel<__int128_t> f1;
-    FinalModel<std::optional<__int128_t>> f2;
-    FinalModel<__uint128_t> f3;
-    FinalModel<std::optional<__uint128_t>> f4;
-};
-
-namespace test {
-
-// Fast Binary Encoding Struct128 final model
-class Struct128FinalModel : public FBE::Model
-{
-public:
-    Struct128FinalModel() : _model(this->buffer(), 8) {}
-    Struct128FinalModel(const std::shared_ptr<FBEBuffer>& buffer) : FBE::Model(buffer), _model(this->buffer(), 8) {}
-
-    // Get the model type
-    static constexpr size_t fbe_type() noexcept { return FinalModel<::test::Struct128>::fbe_type(); }
-
-    // Check if the struct value is valid
-    bool verify();
-
-    // Serialize the struct value
-    size_t serialize(const ::test::Struct128& value);
-    // Deserialize the struct value
-    size_t deserialize(::test::Struct128& value) const noexcept;
-
-    // Move to the next struct value
-    void next(size_t prev) noexcept { _model.fbe_shift(prev); }
-
-private:
-    FinalModel<::test::Struct128> _model;
-};
-
-} // namespace test
-
 } // namespace FBE
