+++ proto/test_pmr_models.h
@@ -1351,99 +1351,4 @@ public:
 
 } // namespace test_pmr
 
-// Fast Binary Encoding ::test_pmr::Struct128 field model
-template <>
-class FieldModel<::test_pmr::Struct128>
-{
-public:
-    FieldModel(FBEBuffer& buffer, size_t offset) noexcept;
-
-    // Get the field offset
-    size_t fbe_offset() const noexcept { return _offset; }
-    // Get the field size
-    size_t fbe_size() const noexcept { return 4; }
-    // Get the field body size
-    size_t fbe_body() const noexcept;
-    // Get the field extra size
-    size_t fbe_extra() const noexcept;
-    // Get the field type
-    static constexpr size_t fbe_type() noexcept { return 144; }
-
-    // Shift the current field offset
-    void fbe_shift(size_t size) noexcept { _offset += size; }
-    // Unshift the current field offset
-    void fbe_unshift(size_t size) noexcept { _offset -= size; }
-
-    // Check if the struct value is valid
-    bool verify(bool fbe_verify_type = true) const noexcept;
-    // Check if the struct fields are valid
-    bool verify_fields(size_t fbe_struct_size) const noexcept;
-
-    // Get the struct value (begin phase)
-    size_t get_begin() const noexcept;
-    // Get the struct value (end phase)
-    void get_end(size_t fbe_begin) const noexcept;
-
-    // Get the struct value
-    void get(::test_pmr::Struct128& fbe_value) const noexcept;
-    // Get the struct fields values
-    void get_fields(::test_pmr::Struct128& fbe_value, size_t fbe_struct_size) const noexcept;
-
-    // Set the struct value (begin phase)
-    size_t set_begin();
-    // Set the struct value (end phase)
-    void set_end(size_t fbe_begin);
-
-    // Set the struct value
-    void set(const ::test_pmr::Struct128& fbe_value) noexcept;
-    // Set the struct fields values
-    void set_fields(const ::test_pmr::Struct128& fbe_value) noexcept;
-
-private:
-    FBEBuffer& _buffer;
-    size_t _offset;
-
-public:
-    FieldModel<__int128_t> f1;
-    FieldModel<std::optional<__int128_t>> f2;
-    FieldModel<__uint128_t> f3;
-    FieldModel<std::optional<__uint128_t>> f4;
-};
-
-namespace test_pmr {
-
-// Fast Binary Encoding Struct128 model
-class Struct128Model : public FBE::Model
-{
-public:
-    Struct128Model() : model(this->buffer(), 4) {}
-    Struct128Model(const std::shared_ptr<FBEBuffer>& buffer) : FBE::Model(buffer), model(this->buffer(), 4) {}
-
-    // Get the model size
-    size_t fbe_size() const noexcept { return model.fbe_size() + model.fbe_extra(); }
-    // Get the model type
-    static constexpr size_t fbe_type() noexcept { return FieldModel<::test_pmr::Struct128>::fbe_type(); }
-
-    // Check if the struct value is valid
-    bool verify();
-
-    // Create a new model (begin phase)
-    size_t create_begin();
-    // Create a new model (end phase)
-    size_t create_end(size_t fbe_begin);
-
-    // Serialize the struct value
-    size_t serialize(const ::test_pmr::Struct128& value);
-    // Deserialize the struct value
-    size_t deserialize(::test_pmr::Struct128& value) const noexcept;
-
-    // Move to the next struct value
-    void next(size_t prev) noexcept { model.fbe_shift(prev); }
-
-public:
-    FieldModel<::test_pmr::Struct128> model;
-};
-
-} // namespace test_pmr
-
 } // namespace FBE
